requirements.txt
# pytorch model related dependencies
torch
torchvision
pillow
colorama
opencv-python
numpy

# tflite model related dependencies
# tensorflow
# onnx
# https://github.com/onnx/onnx-tensorflow/archive/master.zip

# api related dependencies
fastapi
uvicorn
python-multipart
jinja2
aiofiles
rich
